+++ lib/generators/templates/devise.rb
@@ -4,7 +4,8 @@ Devise.setup do |config|
   # The secret key used by Devise. Devise uses this key to generate
   # random tokens. Changing this key will render invalid all existing
   # confirmation, reset password and unlock tokens in the database.
-  config.secret_key = '<%= SecureRandom.hex(64) %>'
+  # Please add it to your 'secrets.yml' file under the 'devise_secret_key' name.
+  # config.secret_key = '<%= SecureRandom.hex(64) %>'
 
   # ==> Mailer Configuration
   # Configure the e-mail address which will be shown in Devise::Mailer,
@@ -95,6 +96,7 @@ Devise.setup do |config|
   config.stretches = Rails.env.test? ? 1 : 10
 
   # Setup a pepper to generate the encrypted password.
+  # Please add it to your 'secrets.yml' file under the 'devise_pepper' name.
   # config.pepper = '<%= SecureRandom.hex(64) %>'
 
   # ==> Configuration for :confirmable